requirements.txt
fastapi
aioinflux
uvicorn
python-decouple
